use bootstrap compilers consistently on osx
the build necessarily needs to compile llvm-tblgen, which requires symbols
from a matching version of llvmdev, which contradicts with the compilers in
the build environment in cross-compilation. Given that, it's easier to also
use the bootstrap compilers on osx-64, in order to be able to specify the
required metadata consistently for osx across both architectures.